Nano pin out
Shows the pin usage and the jmri node address for each pin and also shows the  allocated cmri bit's in my case it also shows the block detection that the signals protect. block detection is current sensors from Merg. the 5 infrared sensors are used more for stopping the trains in python scripts